chore: refactor web-test-runner config
